+++ src/components/radio/radio.ts
@@ -83,11 +83,14 @@ export default class SlRadio extends LitElement {
   }
 
   getAllRadios() {
-    const form = this.closest('sl-form, form') || document.body;
+    const radioGroup = this.closest('sl-radio-group');
 
-    if (!this.name) return [];
+    // Radios must be part of a radio group
+    if (!radioGroup) {
+      return [];
+    }
 
-    return [...form.querySelectorAll('sl-radio')].filter((radio: this) => radio.name === this.name) as this[];
+    return [...radioGroup.querySelectorAll('sl-radio')].filter((radio: this) => radio.name === this.name) as this[];
   }
 
   getSiblingRadios() {
@@ -171,8 +174,8 @@ export default class SlRadio extends LitElement {
             value=${ifDefined(this.value)}
             ?checked=${this.checked}
             ?disabled=${this.disabled}
-            role="radio"
             aria-checked=${this.checked ? 'true' : 'false'}
+            aria-disabled=${this.disabled ? 'true' : 'false'}
             aria-labelledby=${this.labelId}
             @click=${this.handleClick}
             @blur=${this.handleBlur}